Factor integration tests into separate module.
